+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- The buildpack now properly reads Conjur credentials from `VCAP_SERVICES` when
+  `VCAP_SERVICES` contains credentials for other services with the same field
+  names (e.g. `version`).
+
 ## [2.1.2] - 2019-10-28
 
 ### Added